Improve Rust safety checks
